style(docs): fix eslint errors
