+++ apps/docs/src/components/icons/IconCheck.tsx
@@ -10,7 +10,7 @@ export default function IconCheck({ width = 24, height = 24 }: IconCheckProps): 
         <svg
             className="custom-icon"
             width={width}
-            height={width}
+            height={height}
             viewBox="0 0 24 24"
             fill="none"
             xmlns="http://www.w3.org/2000/svg"
